+++ src/main/backend.ts
@@ -14,10 +14,11 @@ import { Dsn, forget, logger, SentryError } from '@sentry/utils';
 import { app, crashReporter, ipcMain } from 'electron';
 import { join } from 'path';
 
-import { CommonBackend, ElectronOptions, getNameFallback, IPC_EVENT, IPC_PING, IPC_SCOPE } from '../common';
+import { CommonBackend, ElectronOptions, getNameFallback, IPC } from '../common';
 import { supportsGetPathCrashDumps, supportsRenderProcessGone } from '../electron-version';
+import { addEventDefaults } from './context';
 import { captureMinidump } from './index';
-import { normalizeUrl } from './normalize';
+import { normalizeEvent, normalizeUrl } from './normalize';
 import { Store } from './store';
 import { NetTransport } from './transports/net';
 import { MinidumpUploader } from './uploader';
@@ -77,6 +78,9 @@ export class MainBackend extends BaseBackend<ElectronOptions> implements CommonB
   /** Uploader for minidump files. */
   private _uploader?: MinidumpUploader;
 
+  /** Counter used to ensure no race condition when updating extra params */
+  private _updateEpoch: number;
+
   /** Creates a new Electron backend instance. */
   public constructor(options: ElectronOptions) {
     // Disable session tracking until we've decided how this should work with Electron
@@ -88,6 +92,7 @@ export class MainBackend extends BaseBackend<ElectronOptions> implements CommonB
     this._scopeStore = new Store<Scope>(getCachePath(), 'scope_v2', new Scope());
     // We need to store the scope in a variable here so it can be attached to minidumps
     this._scopeLastRun = this._scopeStore.get();
+    this._updateEpoch = 0;
 
     this._setupScopeListener();
 
@@ -163,15 +168,94 @@ export class MainBackend extends BaseBackend<ElectronOptions> implements CommonB
     const hubScope = getCurrentHub().getScope();
     if (hubScope) {
       hubScope.addScopeListener(updatedScope => {
-        const cloned = Scope.clone(updatedScope);
+        const scope = Scope.clone(updatedScope);
         // eslint-disable-next-line @typescript-eslint/no-unsafe-member-access
-        (cloned as any)._eventProcessors = [];
+        (scope as any)._eventProcessors = [];
         // eslint-disable-next-line @typescript-eslint/no-unsafe-member-access
-        (cloned as any)._scopeListeners = [];
+        (scope as any)._scopeListeners = [];
 
-        this._scopeStore.set(cloned);
+        this._scopeStore.set(scope);
+
+        // If we use the Crashpad minidump uploader we have to set extra whenever the scope updates
+        //
+        // We do not currently set addExtraParameter on Linux because Breakpad is limited
+        if (this._options.useCrashpadMinidumpUploader === true && process.platform !== 'linux') {
+          this._updateExtraParams(scope);
+        }
       });
     }
+  }
+
+  /** Updates Electron uploader extra params */
+  private _updateExtraParams(scope: Scope): void {
+    this._updateEpoch += 1;
+    const currentEpoch = this._updateEpoch;
+
+    forget(
+      this._getNativeUploaderEvent(scope).then(event => {
+        if (currentEpoch !== this._updateEpoch) return;
+
+        // Update the extra parameters in the main process
+        const mainParams = this._getNativeUploaderExtraParams(event);
+        for (const key of Object.keys(mainParams)) {
+          crashReporter.addExtraParameter(key, mainParams[key]);
+        }
+      }),
+    );
+  }
+
+  /** Builds up an event to send with the native Electron uploader */
+  private async _getNativeUploaderEvent(scope: Scope): Promise<Event> {
+    // Basic SDK info
+    let event: Event = {
+      tags: { event_type: 'native' },
+    };
+
+    // Apply the scope to the event
+    await scope.applyToEvent(event);
+    // Add all the extra context
+    event = await addEventDefaults(this._options.appName, event);
+    return normalizeEvent(event);
+  }
+
+  /** Chunks up event JSON into 1 or more parameters for use with the native Electron uploader
+   *
+   * Returns chunks with keys and values:
+   * {
+   *    sentry__1: '{ json...',
+   *    sentry__2: 'more json...',
+   *    sentry__x: 'end json }',
+   * }
+   */
+  private _getNativeUploaderExtraParams(event: Event): { [key: string]: string } {
+    const maxBytes = 20300;
+
+    /** Max chunk sizes are in bytes so we can't chunk by characters or UTF8 could bite us.
+     *
+     * We attempt to split by space (32) and double quote characters (34) as there are plenty in JSON
+     * and they are guaranteed to not be the first byte of a multi-byte UTF8 character.
+     */
+    let buf = Buffer.from(JSON.stringify(event));
+    const chunks = [];
+    while (buf.length) {
+      // Find last '"'
+      let i = buf.lastIndexOf(34, maxBytes + 1);
+      // Or find last ' '
+      if (i < 0) i = buf.lastIndexOf(32, maxBytes + 1);
+      // Or find first '"'
+      if (i < 0) i = buf.indexOf(34, maxBytes);
+      // Or find first ' '
+      if (i < 0) i = buf.indexOf(32, maxBytes);
+      // We couldn't find any space or quote chars so split at maxBytes and hope for the best 🤷‍♂️
+      if (i < 0) i = maxBytes;
+      chunks.push(buf.slice(0, i + 1).toString());
+      buf = buf.slice(i + 1);
+    }
+
+    return chunks.reduce((acc, cur, i) => {
+      acc[`sentry__${i + 1}`] = cur;
+      return acc;
+    }, {} as { [key: string]: string });
   }
 
   /** Returns whether native reports are enabled. */
@@ -243,11 +327,13 @@ export class MainBackend extends BaseBackend<ElectronOptions> implements CommonB
       contents: Electron.WebContents,
       details?: Electron.RenderProcessGoneDetails,
     ): Promise<void> => {
-      try {
-        await this._sendNativeCrashes(this._getNewEventWithElectronContext(contents, details));
-      } catch (e) {
-        // eslint-disable-next-line no-console
-        console.error(e);
+      if (this._options.useSentryMinidumpUploader !== false) {
+        try {
+          await this._sendNativeCrashes(this._getNewEventWithElectronContext(contents, details));
+        } catch (e) {
+          // eslint-disable-next-line no-console
+          console.error(e);
+        }
       }
 
       addBreadcrumb({
@@ -282,11 +368,11 @@ export class MainBackend extends BaseBackend<ElectronOptions> implements CommonB
 
   /** Installs IPC handlers to receive events and metadata from renderers. */
   private _installIPC(): void {
-    ipcMain.on(IPC_PING, (event: Electron.IpcMainEvent) => {
-      event.sender.send(IPC_PING);
+    ipcMain.on(IPC.PING, (event: Electron.IpcMainEvent) => {
+      event.sender.send(IPC.PING);
     });
 
-    ipcMain.on(IPC_EVENT, (ipc: Electron.IpcMainEvent, jsonEvent: string) => {
+    ipcMain.on(IPC.EVENT, (ipc: Electron.IpcMainEvent, jsonEvent: string) => {
       let event: Event;
       try {
         event = JSON.parse(jsonEvent) as Event;
@@ -304,7 +390,7 @@ export class MainBackend extends BaseBackend<ElectronOptions> implements CommonB
       captureEvent(event);
     });
 
-    ipcMain.on(IPC_SCOPE, (_: any, jsonRendererScope: string) => {
+    ipcMain.on(IPC.SCOPE, (_: any, jsonRendererScope: string) => {
       let rendererScope: Scope;
       try {
         rendererScope = JSON.parse(jsonRendererScope) as Scope;