+++ src/renderer/backend.ts
@@ -3,7 +3,7 @@ import { BaseBackend, getCurrentHub } from '@sentry/core';
 import { Event, EventHint, Scope, Severity } from '@sentry/types';
 import { walk as walkUtil } from '@sentry/utils';
 
-import { CommonBackend, ElectronOptions, getNameFallback, IPC_EVENT, IPC_PING, IPC_SCOPE } from '../common';
+import { CommonBackend, ElectronOptions, getNameFallback, IPC } from '../common';
 import { requiresNativeHandlerRenderer } from '../electron-version';
 
 /** Walks an object to perform a normalization on it with a maximum depth of 50 */
@@ -112,13 +112,13 @@ export class RendererBackend extends BaseBackend<ElectronOptions> implements Com
     const ipcObject = {
       // We pass through JSON because in Electron >= 8, IPC uses v8's structured clone algorithm and throws errors if
       // objects have functions. Calling walk makes sure to break circular references.
-      sendScope: (scope: Scope) => ipcRenderer.send(IPC_SCOPE, JSON.stringify(scope, walk)),
-      sendEvent: (event: Event) => ipcRenderer.send(IPC_EVENT, JSON.stringify(event, walk)),
+      sendScope: (scope: Scope) => ipcRenderer.send(IPC.SCOPE, JSON.stringify(scope, walk)),
+      sendEvent: (event: Event) => ipcRenderer.send(IPC.EVENT, JSON.stringify(event, walk)),
       pingMain: (success: () => void) => {
-        ipcRenderer.once(IPC_PING, () => {
+        ipcRenderer.once(IPC.PING, () => {
           success();
         });
-        ipcRenderer.send(IPC_PING);
+        ipcRenderer.send(IPC.PING);
       },
     };