+++ test/e2e/index.test.ts
@@ -158,7 +158,7 @@ describe('E2E Tests', () => {
           return;
         }
 
-        await context.start('sentry-electron-uploader', 'native-main');
+        await context.start('sentry-electron-uploader-main', 'native-main');
         // It can take rather a long time to get the event on Mac
         await context.waitForEvents(testServer, 1, 20000);
 
@@ -169,6 +169,7 @@ describe('E2E Tests', () => {
         expect(event.method).to.equal('minidump');
 
         if (process.platform !== 'linux') {
+          expect(event.data.user?.id).to.equal('ABCDEF1234567890');
           expect(event.namespaced?.initialScope?.user).to.equal('some_user');
         }
       });
@@ -180,7 +181,7 @@ describe('E2E Tests', () => {
           return;
         }
 
-        await context.start('sentry-electron-uploader', 'native-renderer');
+        await context.start('sentry-electron-uploader-renderer', 'native-renderer');
         // It can take rather a long time to get the event on Mac
         await context.waitForEvents(testServer, 1, 20000);
 
@@ -201,7 +202,7 @@ describe('E2E Tests', () => {
           return;
         }
 
-        await context.start('sentry-electron-uploader', 'native-gpu');
+        await context.start('sentry-electron-uploader-main', 'native-gpu');
         await context.waitForEvents(testServer, 1, 20000);
 
         expect(testServer.events.length).to.equal(1);